+++ dump_cacert.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+pwd=${PWD}
+cd res/cert
+xxd -i cacert.pem > cacert.pem.c 
+cd $pwd