+++ src/base/@base.c
@@ -1,0 +1,13 @@
+// @base.c
+// Module entry point for base.
+// Controls which source files are part of this module build.
+
+
+#include "dir.c"
+#include "file.c"
+#include "list.c"
+#include "log.c"
+#include "memory.c"
+#include "path.c"
+#include "cstr.c"
+#include "url.c"