+++ src/base/@base.h
@@ -1,0 +1,18 @@
+// @base.h
+// Module entry point for base.
+// Controls which source files are part of this module build.
+
+
+
+#include "common.h"
+#include "list.h"
+#include "dir.h"
+#include "file.h"
+#include "log.h"
+#include "memory.h"
+#include "path.h"
+#include "cstr.h"
+#include "url.h"
+
+
+