+++ src/http/@zn_http.c
@@ -1,8 +1,0 @@
-#include "http/ZN_HttpParamValue.c"
-#include "http/ZN_HttpHandleClient.c"
-#include "http/ZN_HttpRequest.c"
-#include "http/ZN_HttpResponse.c"
-#include "http/ZN_HttpServer.c"
-#include "http/ZN_HttpServerMPS.c"
-#include "http/ZN_HttpRoute.c"
-#include "http/ZN_HttpRouteManager.c"