+++ src/zetnet.c
@@ -1,7 +1,8 @@
 #include "zetnet.h"
 
-#include "ZN_TcpServer.c"
-#include "ZN_TcpUtils.c"
+#ifdef __WITH_SSL__
+static SSL_CTX * g_ssl_ctx = NULL;
+#endif
 
 bool ZN_Init(void){
 #ifdef _WIN32
@@ -14,11 +15,55 @@ bool ZN_Init(void){
 	}
 #endif
 
+#ifdef __WITH_SSL__
+    // OpenSSL init
+	if(g_ssl_ctx == NULL){
+		SSL_library_init();
+		SSL_load_error_strings();
+
+		g_ssl_ctx = SSL_CTX_new(TLS_client_method());
+		if (!g_ssl_ctx) {
+			fprintf(stderr,"SSL_CTX_new failed\n");
+			return false;
+		}
+
+		if(SSL_CTX_set_default_verify_paths(g_ssl_ctx) == 0){
+			fprintf(stderr,"SSL_CTX_set_default_verify_paths failed\n");
+			return false;
+		}
+
+
+		if (!ZN_SSL_LoadEmbeddedCA(g_ssl_ctx)) {
+		    fprintf(stderr, "Failed to load embedded CA\n");
+		    return false;
+		}
+
+		SSL_CTX_set_verify(g_ssl_ctx, SSL_VERIFY_PEER, NULL);
+
+	}
+#endif
+
 	return true;
 }
 
+#ifdef __WITH_SSL__
+SSL_CTX * ZN_GetSSLContext(void){
+	return g_ssl_ctx;
+}
+
+#endif
+
 void ZN_DeInit(void){
+#ifdef __WITH_SSL__
+	if(g_ssl_ctx != NULL){
+		SSL_CTX_free(g_ssl_ctx);
+        g_ssl_ctx = NULL;
+	}
+#endif
+
 #ifdef _WIN32
 	WSACleanup();
 #endif
+
+
 }